requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
statsmodels==0.14.0
jupyter==1.0.0
streamlit==1.29.0
plotly==5.18.0
python-dotenv==1.0.0
joblib==1.3.2
